requirements.txt
polars
google-auth
google-api-python-client
streamlit
plotly